feat: support toml config file (#300)
* docs: update readme

* feat(sysproxy): add logging for sysproxy

* fix: parameter type for NewTTLCache

* chore(https): use uint8 for window-size

* feat(dns): enhence loggin for dns package

* feat(config): add support for toml configuration

* fix: condition for patternMatches func

* chore: trim version string

* chore: BurntSushi/toml

* chore(dns): enhence ResolverInfo

* chore: remove trailing period in version string

* feat: reorganize main function to use urfave/cli

* remove(config): unused parse.go

* chore(config): remove unused lines

* chore: let proxy.Start() wait for setup logs to be printed

+++ internal/datastruct/ttlcache.go
@@ -36,7 +36,7 @@ type TTLCache[T any] struct {
 // NewTTLCache creates a new sharded TTL cache with a background janitor goroutine.
 // numShards specifies the number of shards to create and must be greater than 0.
 // cleanupInterval specifies how often the janitor should run.
-func NewTTLCache[T any](numShards uint64, cleanupInterval time.Duration) *TTLCache[T] {
+func NewTTLCache[T any](numShards uint8, cleanupInterval time.Duration) *TTLCache[T] {
 	if numShards == 0 {
 		panic(fmt.Errorf("ttlcache: numShards must be greater than 0, got %d", numShards))
 	}
